requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.1.3
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.4.22
cffi==2.0.0
chardet==7.4.3
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==46.0.7
curl_cffi==0.15.0
dash==2.17.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateutils==0.6.12
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
Flask==3.0.3
frozendict==2.4.7
greenlet==3.4.0
gunicorn==22.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.13
importlib_metadata==9.0.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multitasking==0.0.13
nest-asyncio==1.6.0
nsetools==1.0.11
numpy==1.26.4
orjson==3.11.8
packaging==26.1
pandas==2.2.2
passlib==1.7.4
peewee==4.0.4
pillow==12.2.0
platformdirs==4.9.6
plotly==5.22.0
pluggy==1.6.0
protobuf==7.34.1
psycopg2-binary==2.9.9
pyasn1==0.6.3
pycparser==3.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.20.0
pytest==8.2.0
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.26
pytz==2026.1.post1
PyYAML==6.0.3
redis==5.0.4
reportlab==4.1.0
requests==2.32.2
retrying==1.4.2
rich==15.0.0
rich-toolkit==0.19.7
rsa==4.9.1
ruff==0.4.4
scipy==1.13.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.30
starlette==0.37.2
tenacity==8.2.3
typer==0.24.2
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.30.1
watchfiles==1.1.1
webencodings==0.5.1
websockets==16.0
Werkzeug==3.0.6
wrapt==2.1.2
yfinance==1.3.0
zipp==3.23.1
